build: rename package tools to testsuite
