+++ schematron/v2.2.0/tests/test_v2-2-0_L000_Prelim_Analysis.py
@@ -1,7 +1,7 @@
 import os
 
 
-from tools.validate_sch import validate_schematron
+from testsuite.validate_sch import validate_schematron
 from schematron.conftest import AssertFailureRolesMixin, exemplary_tree, remove_element, v2_2_0_SCH_DIR
