+++ testsuite/constants.py
