+++ tox.ini
@@ -1,26 +1,25 @@
 [tox]
+isolated_build = True
 minversion = 2.4
 envlist =
     python,
     generated_sch
 
 [testenv:python]
-deps=
-    -r{toxinidir}/requirements.txt
+whitelist_externals = poetry
 setenv = TOXINIDIR = {toxinidir}
 passenv =
     HOME
 commands =
-    py.test {posargs}
+    poetry run pytest {posargs}
 extras =
     all
     testing
 
 [testenv:generated_sch]
-deps=
-    -r{toxinidir}/requirements.txt
+whitelist_externals = poetry
 commands =
-    ./buildingsch.py generate_all --dry-run
+    poetry run python ./buildingsch.py generate_all --dry-run
 
 [testenv:pre-commit]
 deps =
